requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
# Brain Tumor dependencies
ultralytics
opencv-python
segment-anything-hq
numpy
pandas
scikit-learn
xgboost
reportlab
bcrypt
pillow
shap
torch
torchvision
google-genai
